docs: Add missing argument in `kubeadm` workaround script.
One of the `sed` commands for updating the kubeadm systemd
file (`/etc/systemd/system/kubelet.service.d/10-kubeadm.conf`)
was missing a `-i` reference to the file itself, causing it
to hang indefinitely if ran as declared.

I also wrapped this second `sed` in an `if grep ...` check,
in order to make the operation idempotent.