charts: Add Core DNS Support (#1176)
* charts: Add Core DNS Support

With CoreDNS becoming a first class citizen it should be an optional part
of the kube-prometheus deployment.

Fixes: #1174

* added kube-prometheus service and service monitor

In order to work directly with a cluster created by bootkube used port 9153,
and created a service exposing metrics since bootkube does does not. Also in
the helm chart changed the default port to 9153 since that is the coredns
plugins default port and to be consistent throughout the repo.