Adding kustomization files for remote bases
Closes #2391

Adds support to allow kustomize users to point to a [remote
base](https://github.com/kubernetes-sigs/kustomize/blob/master/examples/remoteBuild.md) of the prometheus-operator and kube-prometheus. Now that kustomize is a part of kubectl, this is provides a very simple way of installing prometheus-operator and kube-prometheus with no additional tools besides kubectl.

* [Added] kustomization.yaml: hardcoded kustomization base for prometheus-operator
bundle.yaml
* [Added] kustomization generation from kube-prometheus example jsonnet