Changes coming from jb update
When running jb update and make generate the following changes are
unstaged. If I don't add them on a coomit testing pipeline fails.
Adding them in a separate commit us they are out of purpose of my
initial PR. It seems that something gone wrong on a previous commit.
My branch is in sync with upstream/master branch.