rule-sanitizer: avoid sorting rules
This change improves the rule-sanitizer library to avoid sorting all the
rules. This was originally done to simplify indexing rules with the same
names, but it turns out that integrating this lib can be complex to
review when all the rules of the manifests are reordered. Thus, it is
easier to just leave the file as it was initially sorted.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>