blackbox-exporter: fix broken image reference
The currently used version of the configmap-reload image (0.5.0) isn't hosted
on Github (ghcr.io/jimmidyson/configmap-reload), only on
Docker Hub (jimmidyson/configmap-reload), meaning the current reference is broken.
Only the versions from 0.10.0 and onwards are hosted on Github.
The best long term solution would be to bump configmap-reload to a newer
version, but this is something I'm not comfortable doing.