Improve watchdog alert
This expression fires the Watchdog alert also if the TSDB is up to date and therefore checks the functionality of the full stack. With the former `vector(1)`, only alertmanager needs to be functional to fire. So In case of a full TSDB storage, the Watchdog still fires and the lack of new metrics goes unnoticed.