fix many-to-one matching must be explicit error
the left hand side keeps the "status_code" label
in the time series, which can lead to one single
handler having multiple time series. When divided
right hand expression, it won't contain the status_code
leading to many-to-one error.

This is an issue because while the alert is a warning,
a failure to evaluate a rule is a sev1.

I think this has always been incorrect, but I'm
surprised no one has experienced it.

I also noticed the runbook does not really exist,
so I removed it to avoid confusion