main.libsonnet: Align kubePrometheus constructor with other components
All components are constructed by passing `$.values[component]` as a
parameter to the function responsible for initializing it; e.g. the
`prometheus` component is constructed with

```jsonnet
prometheus($.values.prometheus)
```

But the `kubePrometheus` component wasn't following that pattern, making
it hard to pass some configuration options in a standard way.

For instance, before this commit, setting

```jsonnet
{
  values+:: {
    kubePrometheus+: {
      mixin+: {
        _config+:: {
          hostNetworkInterfaceSelector: 'device!~"(lxc|veth).+"',
        },
      },
    },
  },
}
```

would have had no effect.

Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>